Remove tracked vector store artifacts
