+++ my_rag_project/api/__init__.py
@@ -1,0 +1,1 @@
+"""API surfaces for serving the RAG system."""
