+++ my_rag_project/docs/example.txt
