+++ my_rag_project/embeddings/__init__.py
@@ -1,0 +1,1 @@
+"""Embedding utilities and vector stores."""
