+++ my_rag_project/med_vectordata2/.gitkeep
@@ -1,0 +1,1 @@
+This placeholder keeps the local Chroma vector store directory in version control without committing generated databases.
