+++ my_rag_project/mlops/mlflow/README.md
@@ -1,0 +1,3 @@
+# MLflow Artifacts
+
+This directory stores MLflow model metadata for the project structure.
