+++ my_rag_project/mlops/monitor/README.md
@@ -1,0 +1,3 @@
+# Monitoring Configuration
+
+Prometheus and Grafana configuration files live here.
