+++ my_rag_project/models/__init__.py
@@ -1,0 +1,1 @@
+"""Prompt templates and model helpers."""
