+++ my_rag_project/pipelines/__init__.py
@@ -1,0 +1,1 @@
+"""Data ingestion and retraining pipelines."""
