+++ my_rag_project/utils/__init__.py
@@ -1,0 +1,1 @@
+"""General utilities for the RAG project."""
