--- workflows/auto_pipeline.n8n
@@ -1,57 +1,0 @@
-{
-  "name": "Auto Pipeline",
-  "nodes": [
-    {
-      "parameters": {
-        "triggerTimes": [
-          {
-            "mode": "custom",
-            "cronExpression": "0 2 * * *"
-          }
-        ]
-      },
-      "id": "1",
-      "name": "Daily Schedule",
-      "type": "n8n-nodes-base.cron",
-      "typeVersion": 1,
-      "position": [
-        200,
-        300
-      ]
-    },
-    {
-      "parameters": {
-        "command": "dvc repro retrain",
-        "options": {
-          "cwd": "/workspace/tinaColab"
-        }
-      },
-      "id": "2",
-      "name": "Run DVC Pipeline",
-      "type": "n8n-nodes-base.executeCommand",
-      "typeVersion": 1,
-      "position": [
-        500,
-        300
-      ]
-    }
-  ],
-  "connections": {
-    "Daily Schedule": {
-      "main": [
-        [
-          {
-            "node": "Run DVC Pipeline",
-            "type": "main",
-            "index": 0
-          }
-        ]
-      ]
-    }
-  },
-  "settings": {
-    "executionTimeout": 3600
-  },
-  "pinData": {},
-  "active": false
-}
